Use a separate Listener for the subselectpalettes handling.
