+++ src/Resources/config/listeners.yml
@@ -6,3 +6,6 @@ services:
 
   cca.meta_palettes.listener.hook_listener:
     class: ContaoCommunityAlliance\MetaPalettes\Listener\HookListener
+
+  cca.meta_palettes.listener.sub_select_palettes_listener:
+    class: ContaoCommunityAlliance\MetaPalettes\Listener\SubSelectPalettesListener
